Add support for time.Time objects to gorp.
This commit adds support for storing and retriving time.Time objects,
by passing through such objects to the underlying driver.  This
resolves gorp issue #14.

The commit is tested and works with the github.com/mattn/go-sqlite3
driver.

The commit does not currently work with the
github.com/ziutek/mymysql/godrv driver, due to a bug in the mysql
driver.  This problem is resolved by mymysql pull request #77.  I have
successfully tested that with the fix from pull request #77 applied to
mymysql, and with the current commit applied, gorp can correctly store
and retrieve time.Time objects from mysql databases.

This commit is NOT tested with the github.com/lib/pq driver.