Updated Overridden Fields Logic
First, I added breaks in the loops to check for overridden fields when mapping a struct to a table.

Second, I modified the condition to the check to see if a field is considered 'overridden'.  Originally, it was only considered overridden if the column name was the same, but that meant that if you overrode the struct field but added a db tag to set the column name, the overridden field would still be added to the column map (using the field name).  This didn't make much sense.

This updates the conditional to check the field name instead of the column name for overriding purposes, which seems to make more sense.  It could possibly allow more database errors than the prior way of handling this, but I would consider those instances bugs in the underlying code, not in gorp.