Travis Build
I *just* noticed that the travis build was failing because I was trying to make a string ID (an override for an int ID in the test) autoincrement.  Postgres 9.2 on my local box didn't seem to care, but I didn't have mysql to test with, so I'm guessing that causes issues in mysql.