Initialise plans using sync.Once
Previously, plans were initialised through a racy assignment. Now they
are initialised through a sync.Once, which is more clearly race free and
also makes the intent clear (it is initialisation).

With this change it's also necessary to make the plan more pointer-like,
since if you pass a struct around by value with a lock in it, you are
going to have a bad time. This was pointed out to me by the
gometalinter.

Fixes go-gorp/gorp#296.
Supercedes go-gorp/gorp#153.