Answered some questions about test code
There were some questions raised about the Scan method of the PersonValuerScanner in our tests,
so I documented the type in a way that I hope answers the questions.  I am going to try to solve
the discrepency mentioned in those comments, but it doesn't seem easy - the discrepency most
likely comes from the mysql and/or mymysql libraries.