Convert slice arguments in mappers into flat values
As discussed in [1], sometimes the caller wants to use a slice in the mapper
arguments and the driver does not support it. This patch will convert slices
from the mapper into simple values, modifying the SQL to add more placeholders
and the mapper, to put each slice item as an unique entry.

This behaviour is optional and by default will be disabled.

[1] https://github.com/go-gorp/gorp/issues/5