Check error return of rows.Close().
According to
https://github.blog/2020-05-20-three-bugs-in-the-go-mysql-driver/, when
QueryContext is called with a context that is cancelled during scan, you
can receive incomplete or corrupted results. As I understand it, the
corruption is fixed upstream, but it's still possible to get incomplete
results that will only show up in the error result from Close.

It's still possible and correct to call `defer rows.Close()`, since the
database/sql docs say this:

https://godoc.org/database/sql#Rows.Close

> Close is idempotent and does not affect the result of Err.