Update documentation on named bind parameters. (#7)
I wasn't able to find documentation of these in the godoc, so I added
it. Also, the README had out-of-date information. It said named bind
parameters were only supported for Select, but they are also supported
for Exec since 2015 (04b2c714).

The documentation for ExpandSlices was embedded in the DbMap type, which
made it hard to read on godoc - the entire documentation was in one
large `pre` block. I moved it up to the package level docs.

Note: this is a recreation of https://github.com/go-gorp/gorp/pull/421

Co-authored-by: Jacob Hoffman-Andrews <github@hoffman-andrews.com>