Make query logging more extensible
My use-case is that I want to replace placeholders with specific data. If you set custom logger, currently the problem is that logger receives injected data as a string. This is not very interoperable. I don't want to deal with all possible types by parsing stringified arguments. 

This PR passes original arguments to the logger